requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
pillow==11.2.1
reportlab==4.4.2
sqlparse==0.5.3
typing_extensions==4.14.0